Implement navigation and cross-reference support.
Links to LessWrong articles part of the book are now internal to the PDF,
i.e. they can be visited without a network connection, and from within the
PDF viewer. Additionally, a new CSS file 'lw-pdf-paper.css' is introduced
that displays all links as footnotes (in case the document is printed); in
the footnotes, external URLs display normally, but internal links point to
the appropriate page in the book.

Some accompanying changes:

  - an URL map is introduced ('urlmap.json') that lists permanent redirects
    in URLs found in the articles; many articles point to overcomingbias.com
    addresses that ultimately point to articles part of the sequences.

  - the arrow sign ‘↗’ to signal external links is added now from CSS, so
    that it can be dropped in the pdf-paper version.

  - the "id" attribute of articles, for internal navigation, now belongs to
    the span.title element, so that Prince's "target-content" function
    extracts the correct amount of text (rather than the whole article if
    the "id" attribute is in div.article).